Adding excelsheet that has breakdown of Order
Instead of having all the parameters as part of details in the response, we can have smaller objects like User, Route, Address. This will allow developers to easily understand the parameters and provides better readability, instead of having a long list of parameters. Please find the attached excel having the objects , their fields , datatypes  and description